+++ libs/angular-components/tedi/components/buttons/collapse/collapse.stories.ts
@@ -6,8 +6,9 @@ import {
 } from "@storybook/angular";
 
 import { CollapseComponent } from "./collapse.component";
-import { VerticalSpacingDirective } from "@tehik-ee/tedi-angular/tedi";
-import { TextGroupComponent } from "@tehik-ee/tedi-angular/tedi";
+
+import { VerticalSpacingDirective } from "../../../directives/vertical-spacing/vertical-spacing.directive";
+import { TextGroupComponent } from "../../../../community/components/content/text-group/text-group.component";
 
 /**
  * <a href="https://www.figma.com/design/jWiRIXhHRxwVdMSimKX2FF/TEDI-READY-2.0.4-(work-in-progress)?node-id=15433-138256&amp;m=dev" target="_BLANK" class="sbdocs sbdocs-a css-ajfpqr">Figma ↗</a><br>